Move /app actions out to common directory
